Almost got the ellipsoid geometry working.  Almost.
The visible extent business is a bear to do based on the actual
intersection of the camera frustum with the ellipsoid.

Instead, I think a simpler approach where you determine the visible
distance away, and figure how far that is from the camera's lat/lon
would probably work well enough.