Movin' on up...
With the multi-texturing business working, I was free to start
implementing consumption of actual map services.

Got a rudimentary ArcGIS REST API one working now.

Next on the agenda:  Add a second mesh to the basic ellipsoid geometry
to fill in the portions of the globe that lie outside of the visible
extent.  It's really ugly when the user pans onto a blank surface.