Two steps forward, one step back...
Decided to set aside the sphere generation algorithm I'd come up with,
and try a different one.

This new one uses recursion to produce a sort of fractal geometry, which
I suspect is similar to the way that Google Earth handles this problem.

Incidentally, it looks like this approach may actually require fewer
lines of code.