Huge progress today.
Fractal geometry is now about 95 percent working.

Left to do on that are handling of the international date line
(basically, going out of the bounds of the expected extent),
and a couple of apparent edge cases where the standard logic doesn't
work.

Also, altering the logic so it removes out-LODed geometry from the
scene, but keeps it in memory, in case the user comes back, would be
useful.