You can at least see the texture now...
Well, the texture now shows up as something that is obviously a texture,
anyway.
To do list for tomorrow/this weekend:
1.  Add a call somewhere to ask the decorations to load their textures
right after the scene is set up.
2.  implement inheritance of UV coords from the parent detail node when
inheriting the material.
3.  Refine the event loop to eliminate extraneous calls to the heavy
lifting funcs, and to help keep the geometry/textures more in sync
4.  Figure out why it sometimes enters an infinite loop when refreshing
the geometry.
5.  Implement some sort of height map for the detail nodes.  Maybe
gaussian noise or something initially, and then real data later.