Textures working again.  Mostly.
I've got the texture refreshing business working now, and it works
mostly as expected.  In the future, predictive loading would be a good
thing to add, so the user doesn't have to wait for the texture to get
loaded.

The next pressing issue is the apparent endless loop it gets stuck in
when enhancing the geometry in certain areas.  Not sure what that's
about, but it's aggravating.  Also, with the textures, it eventually
runs the system out of memory.  Doesn't bode well for the future, when I
get the height maps working.