endless loop bug (almost) fixed
I think I've figured out the "endless" loop bug.  The deal is, if it has
to enhance a tile where the visible portion is really narrow, you can
undershoot the LOD aggression by several levels.  The solution, I think,
is to impose a check where it looks at the actual extent being enhanced,
and if it WAY overshoots the tile in question on one side, don't
enhance.

Exactly how this needs to work, I'm not sure yet.  I may leave this bug
for later, and instead do the height map thing.

On the subject of the height map, here's how I imagine that working:
Request a map from the map server of the area for each tile.  Instead of
requesting a high resolution map, though, the map is only X by X pixels,
where X is the number of tiles, so there's a pixel for each vertex.
Then, add the value in each pixel to the geometry when generating it.
Yes, I'm aware the event loop may have to be rejiggered a little for
this to work properly.