Various changes...
Fixed a bug that would cause an exception to be thrown if a decoration
returned an undefined texture when asked for one.
Altered the detail node to only request textures from decorations that
claim to provide textures.
Altered the camera controller to accept the  dimensions of the viewport,
and use them to better compute the desired pan distance.
Something else, I think, but I forget now.