Tons of changes...
Got support for the heightmaps about 50 percent working.  It "works,"
sort of, but I think I'm not applying the Z values to the tile vertices
in the correct order or something, as the polys don't show up perfectly
connected in the scene.

Tweaked the detail node geometry generation so that the vertex buffers
are generated in a separate function from the refreshing.  Improves
readability of the code, but also allows support for the heightmap
without recycling/recreating the entire mesh object.