Refined height map business
I lied.  The height map business was still broken.  It's better now, but
still needs some work.  I feel like there ought to be a way to get the
edges to line up with having to gerrimander them on the client side.

Also, need to write in support for the (vastly superior) image service.

Also, improve the camera panning.