Ten days later...
It's been a while.

I got started on the vector geometry layer code.  It's coming along.
I've got it where it displays some (really ugly) dots representing point
features.  Making that a little prettier may be all that's needed for
now, in reality, but it would be nice to make it fully support the
ArcGIS symbology.  Not sure how to do that efficiently, but we'll see.