Rudminetary picking...
I'm working through the process of getting the picking working.  I've
got a scheme that sort of works, but need to finish banging out the
code.

Tonight, I want to try rewriting the ArcGIS layer to use particles in a
mesh, vs. tons of individual sprites.  Hopefully, this will help improve
performance.