Better vector point handling...
I've fixed a problem with the vector point business where, because of
the way that the particle systems work, it wasn't possible to add
features to the existing data.  That's now fixed, so refreshing when the
camera stops moving now works properly.

This involved totally re-writing the code for generating that, and also
may incur a performance hit.