SO close to having this working...
I rejiggered the tiling scheme code a bit, and it's now about 75%
working.  I actually had it more like 90 percent there for a little
while, where it would load full-coverage tiles for the whole world, but
it distorted mid-latitude tiles, so I redid it.

Tip:  Look at the code it's using for deciding which tiles to request.
The whole world should be X by X, but it's requesting X by X-1.