So close I can taste it.
The tiled service is now "working" in the sense that it produces full
coverage of the given texture area.

Additional work is needed to un-project the tiles from mercator to just
plain WGS84.

Also, the code really needs to be optimized in some way so it doesn't
run the browser out of memory so easily.